ref: move tryCapture and subscribeToMintedNft to composable
